+++ dev-tools/reconfigurator-cli/tests/input/cmds-example.txt
@@ -1,3 +1,4 @@
+# This is a legacy test; new tests shouldn't need to set a seed
 load-example --seed test-basic
 load-example --seed test-basic
