+++ dev-tools/reconfigurator-cli/tests/input/cmds-expunge-newly-added.txt
@@ -1,4 +1,5 @@
-load-example
+# This is a legacy test; new tests shouldn't need to set a seed
+load-example --seed test_expunge_newly_added_external_dns
 
 blueprint-show 3f00b694-1b16-4aaa-8f78-e6b3a527b434
 blueprint-edit 3f00b694-1b16-4aaa-8f78-e6b3a527b434 expunge-zone 9995de32-dd52-4eb1-b0eb-141eb84bc739
