+++ dev-tools/reconfigurator-cli/tests/input/cmds-set-zone-images.txt
@@ -1,5 +1,6 @@
 # Load example system
-load-example --nsleds 1 --ndisks-per-sled 3
+# This is a legacy test; new tests shouldn't need to set a seed
+load-example --nsleds 1 --ndisks-per-sled 3 --seed test_set_zone_images
 blueprint-show 971eeb12-1830-4fa0-a699-98ea0164505c
 
 # Set a couple zones' image sources to specific artifacts.
