+++ dev-tools/reconfigurator-cli/tests/input/cmds.txt
@@ -1,3 +1,5 @@
+# This is a legacy test; new tests shouldn't need to set a seed
+set seed test_basic
 sled-list
 inventory-list
 blueprint-list
