+++ dev-tools/reconfigurator-cli/tests/output/cmds-set-mgs-updates-stderr
