+++ dev-tools/reconfigurator-cli/tests/output/cmds-set-zone-images-stderr
