+++ workspace-hack/Cargo.toml
@@ -98,7 +98,7 @@ proc-macro2 = { version = "1.0.94" }
 quote = { version = "1.0.39" }
 rand = { version = "0.8.5", features = ["small_rng"] }
 regex = { version = "1.11.1" }
-regex-automata = { version = "0.4.8", default-features = false, features = ["dfa-onepass", "dfa-search", "hybrid", "meta", "nfa", "perf", "unicode"] }
+regex-automata = { version = "0.4.8", default-features = false, features = ["dfa", "hybrid", "meta", "nfa", "perf", "unicode"] }
 regex-syntax = { version = "0.8.5" }
 reqwest = { git = "https://github.com/seanmonstar/reqwest", rev = "114d97c1a862567b465cb57f14d584626e1cf30c", features = ["blocking", "cookies", "json", "rustls-tls", "stream"] }
 rsa = { version = "0.9.6", features = ["serde", "sha2"] }
@@ -221,7 +221,7 @@ proc-macro2 = { version = "1.0.94" }
 quote = { version = "1.0.39" }
 rand = { version = "0.8.5", features = ["small_rng"] }
 regex = { version = "1.11.1" }
-regex-automata = { version = "0.4.8", default-features = false, features = ["dfa-onepass", "dfa-search", "hybrid", "meta", "nfa", "perf", "unicode"] }
+regex-automata = { version = "0.4.8", default-features = false, features = ["dfa", "hybrid", "meta", "nfa", "perf", "unicode"] }
 regex-syntax = { version = "0.8.5" }
 reqwest = { git = "https://github.com/seanmonstar/reqwest", rev = "114d97c1a862567b465cb57f14d584626e1cf30c", features = ["blocking", "cookies", "json", "rustls-tls", "stream"] }
 rsa = { version = "0.9.6", features = ["serde", "sha2"] }
